+++ nvbench/device_info.cuh
@@ -154,7 +154,7 @@ struct device_info
   [[nodiscard]] std::size_t get_global_memory_bus_bandwidth() const
   { // 2 is for DDR, CHAR_BITS to convert bus_width to bytes.
     return 2 * this->get_global_memory_bus_peak_clock_rate() *
-           (this->get_global_memory_bus_width() / CHAR_BIT);
+           static_cast<std::size_t>(this->get_global_memory_bus_width() / CHAR_BIT);
   }
 
   /// @return The size of the L2 cache in bytes.
